Corrections from PR review
