+++ src/test/java/com/serenitydojo/fruitmarket/ShoppingCartTest.java
@@ -2,17 +2,22 @@ package com.serenitydojo.fruitmarket;
 
 import com.serenitydojo.Cart;
 import com.serenitydojo.Catalog;
-import io.cucumber.java.Before;
+import io.cucumber.java.ja.但し;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
+@DisplayName("The Shopping Cart")
 
-public class TheShoppingCart {
+public class ShoppingCartTest {
+
+
     Catalog catalog = new Catalog();
+
     @BeforeEach
-    public void setUpCatalog(){
+    public void setUpCatalog() {
         catalog.setPrice("Apples", 4.00);
         catalog.setPrice("Oranges", 5.50);
         catalog.setPrice("Bananas", 6.00);
@@ -21,7 +26,7 @@ public class TheShoppingCart {
 
     @DisplayName("Should initially be empty")
     @Test
-    public void shouldBeEmptyWhenCreated(){
+    public void shouldBeEmptyWhenCreated() {
         Cart cart = new Cart(catalog);
         int numberOfItems = cart.getNumberOfItems();
         assertThat(numberOfItems).isEqualTo(0);
@@ -29,7 +34,7 @@ public class TheShoppingCart {
 
     @DisplayName("Should be added with a single item in the cart")
     @Test
-    public void shouldBeAbleToAddAnItem(){
+    public void shouldBeAbleToAddAnItem() {
         Cart cart = new Cart(catalog);
         cart.addToCart("Apples", 2.0);
         double numberOfItems = cart.getNumberOfItems();
@@ -37,10 +42,9 @@ public class TheShoppingCart {
     }
 
 
-
     @DisplayName("Should keep a running total of the items in the cart.")
     @Test
-    public void shouldKeepARunningTotal(){
+    public void shouldKeepARunningTotal() {
         Cart cart = new Cart(catalog);
         cart.addToCart("Apples", 2.0);
         cart.addToCart("Bananas", 1.0);
@@ -52,7 +56,7 @@ public class TheShoppingCart {
 
     @DisplayName("Should give a 10% discount when you buy 5 kilos of a fruit")
     @Test
-    public void shouldApplyDiscountWhenBuying5KilosOrMore(){
+    public void shouldApplyDiscountWhenBuying5KilosOrMore() {
         Cart cart = new Cart(catalog);
         cart.addToCart("Apples", 5.0);
 
@@ -64,7 +68,7 @@ public class TheShoppingCart {
     @DisplayName("Should give a 10% discount when you buy 5 kilos or more with different fruits")
 
     @Test
-    public void shouldApplyDiscountWhenBuying5KilosOrMoreWithDifferentFruits(){
+    public void shouldApplyDiscountWhenBuying5KilosOrMoreWithDifferentFruits() {
         Cart cart = new Cart(catalog);
         cart.addToCart("Apples", 5.0);
         cart.addToCart("Bananas", 1.0);
